+++ CHANGELOG.md
@@ -54,5 +54,6 @@
     - [Fool's Mate](https://en.wikipedia.org/wiki/Fool's_mate)
 - [Minimax](https://en.wikipedia.org/wiki/Minimax)
 - Adding the AI to the GUI
+- Fixes to the taken pieces icons absolute path
 
 <!-- [1.0.0]: https://github.com/scriptjumper/Java-Chess-Engine/releases/tag/v-1.0.0 -->